Fix synchronization of go channels and cleanup [VPNAND-1913]
The old code in SetConnectivity (and other similar methods) checks the
value of conn.closed to see if it's ok to send to the
conn.updateConnectivity channel. If there are several calls the queue
up on the send operation, after the conn.closed check. If the connection
is closed the pending send can happen after the channel is closed.
(the cleanup method tries to read all the data from channels but it's
not guaranteed to do so).
The new code uses the "select" pattern for channel cancellation and
WaitGroup to wait for everything to finish.