LocalAgent: fix send after closed.
Channels used by SetFeatures, SendGetStatus, SetConnectivity should not
be closed to avoid sending to them after close.
All reads on the channels are done in a safe way (mostly with select
together with closeChannel) so that reads will not get stuck when
closeChannel is closed.
The channels will later be garbage collected, there is no need to close
them. Values sent to them when AgentConnection is closing will be lost,
which is ok.